+++ WiselySplitServer/src/main/java/ca/mohawkCollege/WiselySplitServer/models/Invite.java
@@ -3,11 +3,11 @@ package ca.mohawkCollege.wiselySplitServer.models;
 import java.sql.Timestamp;
 
 public class Invite {
-    private int inviteId;
-    private int senderId;
-    private Integer receiverId;
+    private long inviteId;
+    private long senderId;
+    private Long receiverId;
     private String receiverEmail;
-    private Integer groupId;
+    private Long groupId;
     private String type;   // USER or GROUP
     private String status; // PENDING, ACCEPTED, REJECTED, EXPIRED
     private Timestamp createdAt;
@@ -15,20 +15,20 @@ public class Invite {
 
     public Invite() {}
 
-    public int getInviteId() { return inviteId; }
-    public void setInviteId(int inviteId) { this.inviteId = inviteId; }
+    public long getInviteId() { return inviteId; }
+    public void setInviteId(long inviteId) { this.inviteId = inviteId; }
 
-    public int getSenderId() { return senderId; }
-    public void setSenderId(int senderId) { this.senderId = senderId; }
+    public long getSenderId() { return senderId; }
+    public void setSenderId(long senderId) { this.senderId = senderId; }
 
-    public Integer getReceiverId() { return receiverId; }
-    public void setReceiverId(Integer receiverId) { this.receiverId = receiverId; }
+    public Long getReceiverId() { return receiverId; }
+    public void setReceiverId(Long receiverId) { this.receiverId = receiverId; }
 
     public String getReceiverEmail() { return receiverEmail; }
     public void setReceiverEmail(String receiverEmail) { this.receiverEmail = receiverEmail; }
 
-    public Integer getGroupId() { return groupId; }
-    public void setGroupId(Integer groupId) { this.groupId = groupId; }
+    public Long getGroupId() { return groupId; }
+    public void setGroupId(Long groupId) { this.groupId = groupId; }
 
     public String getType() { return type; }
     public void setType(String type) { this.type = type; }
